+++ src/Monad.ts
@@ -1,13 +1,12 @@
 /**
- * The `Monad` type class combines the operations of the `Chain` and
- * `Applicative` type classes. Therefore, `Monad` instances represent type
- * constructors which support sequential composition, and also lifting of
- * functions of arbitrary arity.
+ * The `Monad` type class extends the `Functor` type class with a `chain` operation which composes computations in
+ * sequence, using the return value of one computation to determine the next computation.
  *
- * Instances must satisfy the following laws in addition to the `Applicative` and `Chain` laws:
+ * Instances must satisfy the following law in addition to the `Functor` laws:
  *
- * 1. Left identity: `M.chain(M.of(a), f) = f(a)`
- * 2. Right identity: `M.chain(fa, M.of) = fa`
+ * 1. Left identity: `M.chain(M.of(a), f) <-> f(a)`
+ * 2. Right identity: `M.chain(fa, M.of) <-> fa`
+ * 3. Associativity: `M.chain(M.chain(fa, afb), bfc) <-> M.chain(fa, a => M.chain(afb(a), bfc))`
  *
  * Note. `Functor`'s `map` can be derived: `A.map = (fa, f) => A.chain(fa, a => A.of(f(a)))`
  *
@@ -22,40 +21,55 @@ import {
   Applicative4,
   Applicative3C
 } from './Applicative'
-import { Chain, Chain1, Chain2, Chain2C, Chain3, Chain4, Chain3C } from './Chain'
-import { URIS, URIS2, URIS3, URIS4 } from './HKT'
+import { URIS, URIS2, URIS3, URIS4, HKT, Kind, Kind2, Kind3, Kind4 } from './HKT'
 
 /**
  * @since 2.0.0
  */
-export interface Monad<F> extends Applicative<F>, Chain<F> {}
+export interface Monad<F> extends Applicative<F> {
+  readonly chain: <A, B>(f: (a: A) => HKT<F, B>) => (fa: HKT<F, A>) => HKT<F, B>
+}
 
 /**
  * @since 2.0.0
  */
-export interface Monad1<F extends URIS> extends Applicative1<F>, Chain1<F> {}
+export interface Monad1<F extends URIS> extends Applicative1<F> {
+  readonly chain: <A, B>(f: (a: A) => Kind<F, B>) => (fa: Kind<F, A>) => Kind<F, B>
+}
 
 /**
  * @since 2.0.0
  */
-export interface Monad2<M extends URIS2> extends Applicative2<M>, Chain2<M> {}
+export interface Monad2<M extends URIS2> extends Applicative2<M> {
+  readonly chain: <A, E, B>(f: (a: A) => Kind2<M, E, B>) => (fa: Kind2<M, E, A>) => Kind2<M, E, B>
+}
 
 /**
  * @since 2.0.0
  */
-export interface Monad2C<M extends URIS2, L> extends Applicative2C<M, L>, Chain2C<M, L> {}
+export interface Monad2C<M extends URIS2, E> extends Applicative2C<M, E> {
+  readonly chain: <A, B>(f: (a: A) => Kind2<M, E, B>) => (fa: Kind2<M, E, A>) => Kind2<M, E, B>
+}
 
 /**
  * @since 2.0.0
  */
-export interface Monad3<M extends URIS3> extends Applicative3<M>, Chain3<M> {}
+export interface Monad3<M extends URIS3> extends Applicative3<M> {
+  readonly chain: <A, R, E, B>(f: (a: A) => Kind3<M, R, E, B>) => (fa: Kind3<M, R, E, A>) => Kind3<M, R, E, B>
+}
 
 /**
  * @since 2.2.0
  */
-export interface Monad3C<M extends URIS3, E> extends Applicative3C<M, E>, Chain3C<M, E> {}
+export interface Monad3C<M extends URIS3, E> extends Applicative3C<M, E> {
+  readonly chain: <A, R, B>(f: (a: A) => Kind3<M, R, E, B>) => (fa: Kind3<M, R, E, A>) => Kind3<M, R, E, B>
+}
 
 /**
  * @since 2.0.0
  */
-export interface Monad4<M extends URIS4> extends Applicative4<M>, Chain4<M> {}
+export interface Monad4<M extends URIS4> extends Applicative4<M> {
+  readonly chain: <A, S, R, E, B>(
+    f: (a: A) => Kind4<M, S, R, E, B>
+  ) => (fa: Kind4<M, S, R, E, A>) => Kind4<M, S, R, E, B>
+}