+++ src/ReadonlyTuple.ts
@@ -4,7 +4,6 @@
 import { Applicative, Applicative2C } from './Applicative'
 import { Apply2C } from './Apply'
 import { Bifunctor2 } from './Bifunctor'
-import { Chain2C } from './Chain'
 import { Comonad2 } from './Comonad'
 import { Foldable2 } from './Foldable'
 import * as F from './function'
@@ -86,8 +85,8 @@ export function getApplicative<S>(M: Monoid<S>): Applicative2C<URI, S> {
 /**
  * @since 2.5.0
  */
-export function getChain<S>(S: Semigroup<S>): Chain2C<URI, S> {
-  const A = getApply(S)
+export function getMonad<S>(M: Monoid<S>): Monad2C<URI, S> {
+  const A = getApply(M)
   return {
     URI,
     _E: A._E,
@@ -95,22 +94,8 @@ export function getChain<S>(S: Semigroup<S>): Chain2C<URI, S> {
     ap: A.ap,
     chain: (f) => (ma) => {
       const [b, s] = f(fst(ma))
-      return [b, S.concat(snd(ma), s)]
-    }
-  }
-}
-
-/**
- * @since 2.5.0
- */
-export function getMonad<S>(M: Monoid<S>): Monad2C<URI, S> {
-  const C = getChain(M)
-  return {
-    URI,
-    _E: C._E,
-    map: C.map,
-    ap: C.ap,
-    chain: C.chain,
+      return [b, M.concat(snd(ma), s)]
+    },
     of: of(M)
   }
 }