+++ src/index.ts
@@ -15,7 +15,6 @@ import * as boundedJoinSemilattice from './BoundedJoinSemilattice'
 import * as boundedLattice from './BoundedLattice'
 import * as boundedMeetSemilattice from './BoundedMeetSemilattice'
 import * as category from './Category'
-import * as chain from './Chain'
 import * as comonad from './Comonad'
 import * as compactable from './Compactable'
 import * as console from './Console'
@@ -140,10 +139,6 @@ export {
    * @since 2.0.0
    */
   category,
-  /**
-   * @since 2.0.0
-   */
-  chain,
   /**
    * @since 2.0.0
    */